+++ pkg/router/error.go
@@ -1,0 +1,7 @@
+package router
+
+import "errors"
+
+var (
+	ErrAlreadyExists = errors.New("already exists")
+)